Add header files for all behavior tree plugins
