+++ nav2_behavior_tree/include/nav2_behavior_tree/plugins/action/reinitialize_global_localization_service.hpp
@@ -12,24 +12,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#ifndef NAV2_BEHAVIOR_TREE__PLUGINS__ACTION__REINITIALIZE_GLOBAL_LOCALIZATION_SERVICE_HPP_
+#define NAV2_BEHAVIOR_TREE__PLUGINS__ACTION__REINITIALIZE_GLOBAL_LOCALIZATION_SERVICE_HPP_
+
 #include <string>
 
-#include "nav2_behavior_tree/plugins/action/clear_costmap_service.hpp"
+#include "nav2_behavior_tree/bt_service_node.hpp"
+#include "std_srvs/srv/empty.hpp"
 
 namespace nav2_behavior_tree
 {
 
-ClearEntireCostmapService::ClearEntireCostmapService(
-  const std::string & service_node_name,
-  const BT::NodeConfiguration & conf)
-: BtServiceNode<nav2_msgs::srv::ClearEntireCostmap>(service_node_name, conf)
+class ReinitializeGlobalLocalizationService : public BtServiceNode<std_srvs::srv::Empty>
 {
-}
+public:
+  ReinitializeGlobalLocalizationService(
+    const std::string & service_node_name,
+    const BT::NodeConfiguration & conf);
+};
 
 }  // namespace nav2_behavior_tree
 
-#include "behaviortree_cpp_v3/bt_factory.h"
-BT_REGISTER_NODES(factory)
-{
-  factory.registerNodeType<nav2_behavior_tree::ClearEntireCostmapService>("ClearEntireCostmap");
-}
+#endif  // NAV2_BEHAVIOR_TREE__PLUGINS__ACTION__REINITIALIZE_GLOBAL_LOCALIZATION_SERVICE_HPP_
