+++ nav2_behavior_tree/include/nav2_behavior_tree/plugins/condition/initial_pose_received_condition.hpp
@@ -12,23 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "nav2_behavior_tree/plugins/condition/initial_pose_received_condition.hpp"
+#ifndef NAV2_BEHAVIOR_TREE__PLUGINS__CONDITION__INITIAL_POSE_RECEIVED_CONDITION_HPP_
+#define NAV2_BEHAVIOR_TREE__PLUGINS__CONDITION__INITIAL_POSE_RECEIVED_CONDITION_HPP_
+
+#include "behaviortree_cpp_v3/behavior_tree.h"
 
 namespace nav2_behavior_tree
 {
-
-BT::NodeStatus initialPoseReceived(BT::TreeNode & tree_node)
-{
-  auto initPoseReceived = tree_node.config().blackboard->get<bool>("initial_pose_received");
-  return initPoseReceived ? BT::NodeStatus::SUCCESS : BT::NodeStatus::FAILURE;
+BT::NodeStatus initialPoseReceived(BT::TreeNode & tree_node);
 }
 
-}  // namespace nav2_behavior_tree
-
-#include "behaviortree_cpp_v3/bt_factory.h"
-BT_REGISTER_NODES(factory)
-{
-  factory.registerSimpleCondition(
-    "InitialPoseReceived",
-    std::bind(&nav2_behavior_tree::initialPoseReceived, std::placeholders::_1));
-}
+#endif  // NAV2_BEHAVIOR_TREE__PLUGINS__CONDITION__INITIAL_POSE_RECEIVED_CONDITION_HPP_
