+++ nav2_behavior_tree/plugins/action/back_up_action.cpp
@@ -12,59 +12,37 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef NAV2_BEHAVIOR_TREE__BACK_UP_ACTION_HPP_
-#define NAV2_BEHAVIOR_TREE__BACK_UP_ACTION_HPP_
-
 #include <string>
 #include <memory>
 #include <cmath>
 
-#include "nav2_behavior_tree/bt_action_node.hpp"
-#include "nav2_msgs/action/back_up.hpp"
+#include "nav2_behavior_tree/plugins/action/back_up_action.hpp"
 
 namespace nav2_behavior_tree
 {
 
-class BackUpAction : public BtActionNode<nav2_msgs::action::BackUp>
+BackUpAction::BackUpAction(
+  const std::string & xml_tag_name,
+  const std::string & action_name,
+  const BT::NodeConfiguration & conf)
+: BtActionNode<nav2_msgs::action::BackUp>(xml_tag_name, action_name, conf)
 {
-public:
-  BackUpAction(
-    const std::string & xml_tag_name,
-    const std::string & action_name,
-    const BT::NodeConfiguration & conf)
-  : BtActionNode<nav2_msgs::action::BackUp>(xml_tag_name, action_name, conf)
-  {
-    double dist;
-    getInput("backup_dist", dist);
-    double speed;
-    getInput("backup_speed", speed);
+  double dist;
+  getInput("backup_dist", dist);
+  double speed;
+  getInput("backup_speed", speed);
 
-    // silently fix, vector direction determined by distance sign
-    if (speed < 0.0) {
-      speed *= -1.0;
-    }
-
-    // Populate the input message
-    goal_.target.x = dist;
-    goal_.target.y = 0.0;
-    goal_.target.z = 0.0;
-    goal_.speed = speed;
+  // silently fix, vector direction determined by distance sign
+  if (speed < 0.0) {
+    speed *= -1.0;
   }
 
-  void on_tick() override
-  {
-    increment_recovery_count();
-  }
-
-  static BT::PortsList providedPorts()
-  {
-    return providedBasicPorts(
-      {
-        BT::InputPort<double>("backup_dist", -0.15, "Distance to backup"),
-        BT::InputPort<double>("backup_speed", 0.025, "Speed at which to backup")
-      });
-  }
-};
+  // Populate the input message
+  goal_.target.x = dist;
+  goal_.target.y = 0.0;
+  goal_.target.z = 0.0;
+  goal_.speed = speed;
+}
 
 }  // namespace nav2_behavior_tree
 
@@ -80,5 +58,3 @@ BT_REGISTER_NODES(factory)
 
   factory.registerBuilder<nav2_behavior_tree::BackUpAction>("BackUp", builder);
 }
-
-#endif  // NAV2_BEHAVIOR_TREE__BACK_UP_ACTION_HPP_
