+++ nav2_behavior_tree/plugins/action/follow_path_action.cpp
@@ -21,50 +21,39 @@
 #include "nav2_msgs/action/follow_path.hpp"
 #include "nav2_behavior_tree/bt_action_node.hpp"
 
+#include "nav2_behavior_tree/plugins/action/follow_path_action.hpp"
+
 namespace nav2_behavior_tree
 {
 
-class FollowPathAction : public BtActionNode<nav2_msgs::action::FollowPath>
+FollowPathAction::FollowPathAction(
+  const std::string & xml_tag_name,
+  const std::string & action_name,
+  const BT::NodeConfiguration & conf)
+: BtActionNode<nav2_msgs::action::FollowPath>(xml_tag_name, action_name, conf)
 {
-public:
-  FollowPathAction(
-    const std::string & xml_tag_name,
-    const std::string & action_name,
-    const BT::NodeConfiguration & conf)
-  : BtActionNode<nav2_msgs::action::FollowPath>(xml_tag_name, action_name, conf)
-  {
+  config().blackboard->set("path_updated", false);
+}
+
+void FollowPathAction::on_tick()
+{
+  getInput("path", goal_.path);
+  getInput("controller_id", goal_.controller_id);
+}
+
+void FollowPathAction::on_wait_for_result()
+{
+  // Check if the goal has been updated
+  if (config().blackboard->get<bool>("path_updated")) {
+    // Reset the flag in the blackboard
     config().blackboard->set("path_updated", false);
-  }
 
-  void on_tick() override
-  {
+    // Grab the new goal and set the flag so that we send the new goal to
+    // the action server on the next loop iteration
     getInput("path", goal_.path);
-    getInput("controller_id", goal_.controller_id);
+    goal_updated_ = true;
   }
-
-  void on_wait_for_result() override
-  {
-    // Check if the goal has been updated
-    if (config().blackboard->get<bool>("path_updated")) {
-      // Reset the flag in the blackboard
-      config().blackboard->set("path_updated", false);
-
-      // Grab the new goal and set the flag so that we send the new goal to
-      // the action server on the next loop iteration
-      getInput("path", goal_.path);
-      goal_updated_ = true;
-    }
-  }
-
-  static BT::PortsList providedPorts()
-  {
-    return providedBasicPorts(
-      {
-        BT::InputPort<nav_msgs::msg::Path>("path", "Path to follow"),
-        BT::InputPort<std::string>("controller_id", ""),
-      });
-  }
-};
+}
 
 }  // namespace nav2_behavior_tree
