+++ nav2_behavior_tree/plugins/action/reinitialize_global_localization_service.cpp
@@ -17,24 +17,18 @@
 
 #include <string>
 #include <memory>
-#include <cmath>
 
-#include "nav2_behavior_tree/bt_service_node.hpp"
-#include "std_srvs/srv/empty.hpp"
+#include "nav2_behavior_tree/plugins/action/reinitialize_global_localization_service.hpp"
 
 namespace nav2_behavior_tree
 {
 
-class ReinitializeGlobalLocalizationService : public BtServiceNode<std_srvs::srv::Empty>
+ReinitializeGlobalLocalizationService::ReinitializeGlobalLocalizationService(
+  const std::string & service_node_name,
+  const BT::NodeConfiguration & conf)
+: BtServiceNode<std_srvs::srv::Empty>(service_node_name, conf)
 {
-public:
-  ReinitializeGlobalLocalizationService(
-    const std::string & service_node_name,
-    const BT::NodeConfiguration & conf)
-  : BtServiceNode<std_srvs::srv::Empty>(service_node_name, conf)
-  {
-  }
-};
+}
 
 }  // namespace nav2_behavior_tree
