+++ nav2_behavior_tree/plugins/action/spin_action.cpp
@@ -12,49 +12,29 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef NAV2_BEHAVIOR_TREE__SPIN_ACTION_HPP_
-#define NAV2_BEHAVIOR_TREE__SPIN_ACTION_HPP_
-
 #include <string>
 #include <memory>
-#include <cmath>
 
-#include "nav2_behavior_tree/bt_action_node.hpp"
-#include "nav2_msgs/action/spin.hpp"
-#include "geometry_msgs/msg/quaternion.hpp"
-#include "tf2/LinearMath/Quaternion.h"
-#include "tf2_geometry_msgs/tf2_geometry_msgs.h"
+#include "nav2_behavior_tree/plugins/action/spin_action.hpp"
 
 namespace nav2_behavior_tree
 {
 
-class SpinAction : public BtActionNode<nav2_msgs::action::Spin>
+SpinAction::SpinAction(
+  const std::string & xml_tag_name,
+  const std::string & action_name,
+  const BT::NodeConfiguration & conf)
+: BtActionNode<nav2_msgs::action::Spin>(xml_tag_name, action_name, conf)
 {
-public:
-  SpinAction(
-    const std::string & xml_tag_name,
-    const std::string & action_name,
-    const BT::NodeConfiguration & conf)
-  : BtActionNode<nav2_msgs::action::Spin>(xml_tag_name, action_name, conf)
-  {
-    double dist;
-    getInput("spin_dist", dist);
-    goal_.target_yaw = dist;
-  }
+  double dist;
+  getInput("spin_dist", dist);
+  goal_.target_yaw = dist;
+}
 
-  void on_tick() override
-  {
-    increment_recovery_count();
-  }
-
-  static BT::PortsList providedPorts()
-  {
-    return providedBasicPorts(
-      {
-        BT::InputPort<double>("spin_dist", 1.57, "Spin distance")
-      });
-  }
-};
+void SpinAction::on_tick()
+{
+  increment_recovery_count();
+}
 
 }  // namespace nav2_behavior_tree
 
@@ -69,5 +49,3 @@ BT_REGISTER_NODES(factory)
 
   factory.registerBuilder<nav2_behavior_tree::SpinAction>("Spin", builder);
 }
-
-#endif  // NAV2_BEHAVIOR_TREE__SPIN_ACTION_HPP_
