+++ nav2_behavior_tree/plugins/condition/time_expired_condition.cpp
@@ -13,15 +13,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef NAV2_BEHAVIOR_TREE__TIME_EXPIRED_CONDITION_HPP_
-#define NAV2_BEHAVIOR_TREE__TIME_EXPIRED_CONDITION_HPP_
-
 #include <string>
 #include <memory>
 
 #include "behaviortree_cpp_v3/condition_node.h"
 
-#include "nav2_behavior_tree/plugins/time_expired_condition.hpp"
+#include "nav2_behavior_tree/plugins/condition/time_expired_condition.hpp"
 
 namespace nav2_behavior_tree
 {
@@ -65,5 +62,3 @@ BT_REGISTER_NODES(factory)
 {
   factory.registerNodeType<nav2_behavior_tree::TimeExpiredCondition>("TimeExpired");
 }
-
-#endif  // NAV2_BEHAVIOR_TREE__TIME_EXPIRED_CONDITION_HPP_
