+++ nav2_behavior_tree/plugins/control/round_robin_node.cpp
@@ -14,65 +14,56 @@
 
 #include <string>
 
-#include "behaviortree_cpp_v3/control_node.h"
-#include "behaviortree_cpp_v3/bt_factory.h"
+#include "nav2_behavior_tree/plugins/control/round_robin_node.hpp"
 
 namespace nav2_behavior_tree
 {
 
-class RoundRobinNode : public BT::ControlNode
+RoundRobinNode::RoundRobinNode(const std::string & name)
+: BT::ControlNode::ControlNode(name, {})
 {
-public:
-  explicit RoundRobinNode(const std::string & name)
-  : BT::ControlNode::ControlNode(name, {})
-  {
-    setRegistrationID("RoundRobin");
+  setRegistrationID("RoundRobin");
+}
+
+BT::NodeStatus RoundRobinNode::tick()
+{
+  const unsigned num_children = children_nodes_.size();
+
+  setStatus(BT::NodeStatus::RUNNING);
+
+  TreeNode * child_node = children_nodes_[current_child_idx_];
+  const BT::NodeStatus child_status = child_node->executeTick();
+
+  switch (child_status) {
+    case BT::NodeStatus::SUCCESS:
+      // Wrap around to the first child
+      if (++current_child_idx_ == num_children) {
+        // TODO(mjeronimo): halt this child (not all children)
+        current_child_idx_ = 0;
+      }
+
+      ControlNode::haltChildren();
+      return BT::NodeStatus::SUCCESS;
+
+    case BT::NodeStatus::FAILURE:
+      ControlNode::haltChildren();
+      return BT::NodeStatus::FAILURE;
+
+    case BT::NodeStatus::RUNNING:
+      break;
+
+    default:
+      throw BT::LogicError("Invalid status return from BT node");
   }
 
-  BT::NodeStatus tick() override
-  {
-    const unsigned num_children = children_nodes_.size();
+  return BT::NodeStatus::RUNNING;
+}
 
-    setStatus(BT::NodeStatus::RUNNING);
-
-    TreeNode * child_node = children_nodes_[current_child_idx_];
-    const BT::NodeStatus child_status = child_node->executeTick();
-
-    switch (child_status) {
-      case BT::NodeStatus::SUCCESS:
-        // Wrap around to the first child
-        if (++current_child_idx_ == num_children) {
-          // TODO(mjeronimo): halt this child (not all children)
-          current_child_idx_ = 0;
-        }
-
-        ControlNode::haltChildren();
-        return BT::NodeStatus::SUCCESS;
-
-      case BT::NodeStatus::FAILURE:
-        ControlNode::haltChildren();
-        return BT::NodeStatus::FAILURE;
-
-      case BT::NodeStatus::RUNNING:
-        break;
-
-      default:
-        throw BT::LogicError("Invalid status return from BT node");
-        break;
-    }
-
-    return BT::NodeStatus::RUNNING;
-  }
-
-  void halt() override
-  {
-    ControlNode::halt();
-    current_child_idx_ = 0;
-  }
-
-private:
-  unsigned int current_child_idx_{0};
-};
+void RoundRobinNode::halt()
+{
+  ControlNode::halt();
+  current_child_idx_ = 0;
+}
 
 }  // namespace nav2_behavior_tree
